+++ Examples/ObjC/MoEObjCSample/AppDelegate.m
@@ -10,7 +10,7 @@
 #import <MoEngage/MoEngage.h>
 
 
-@interface AppDelegate () <UNUserNotificationCenterDelegate>
+@interface AppDelegate () <UNUserNotificationCenterDelegate, MOMessagingDelegate>
 
 @end
 
@@ -19,6 +19,7 @@
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
     
+    MOMessaging.sharedInstance.messagingDelegate = self;
     //TODO: Add your App Group ID
     [MoEngage setAppGroupID:@"Your App Group ID"];
     
@@ -115,7 +116,24 @@ didReceiveNotificationResponse:(UNNotificationResponse *)response
     }
 }
 
+-(void)notificationClickedWithScreenName:(NSString *)screenName andKVPairs:(NSDictionary *)kvPairs{
+    if (screenName) {
+        NSLog(@"Screen Name : %@",screenName);
+    }
+    if (kvPairs) {
+        NSLog(@"KV Pairs : %@",kvPairs);
+    }
+}
 
+-(void)notificationClickedWithScreenName:(NSString *)screenName KVPairs:(NSDictionary *)kvPairs andPushPayload:(NSDictionary *)userInfo{
+    NSLog(@"Push Payload: %@",userInfo);
+    if (screenName) {
+        NSLog(@"Screen Name : %@",screenName);
+    }
+    if (kvPairs) {
+        NSLog(@"KV Pairs : %@",kvPairs);
+    }
+}
 
 
 @end