+++ Frameworks/MORichNotification.framework/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log:
 
+### Version 4.1.0  *(17th December, 2020)*
+-------------------------------------------
+* Changes to support  `enableSDK()` and `disableSDK()` methods for blocking/unblocking SDK features
+
 ### Version 4.0.1  *(2nd November, 2020)*
 -------------------------------------------
 * Xcode 12 Build settings changes to exclude arm64 architecture for Simulator.